machineLearningServer/requirements.txt
flask
flask_cors
bcrypt
python-dotenv 
opencv-python